@@ -3,8 +3,8 @@ plugins {
 }
 
 // Java compatability
-sourceCompatibility = 1.6
-targetCompatibility = 1.6
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 group 'io.deepstream'
 version '2.2.2'
@@ -28,13 +28,10 @@ task apidoc(type: Javadoc) {
 }
 
 dependencies {
-    compile 'com.google.code.gson:gson:2.6.2'
-    compile 'org.java-websocket:Java-WebSocket:1.3.5'
+    compile 'com.google.code.gson:gson:2.8.6'
+    compile 'org.java-websocket:Java-WebSocket:1.5.1'
     compile 'com.google.j2objc:j2objc-annotations:1.1'
 
     testCompile group: 'junit', name: 'junit', version: '4.11'
     testCompile "org.mockito:mockito-core:1.+"
-
-    sourceCompatibility = 1.6
-    targetCompatibility = 1.6
 }
