@@ -1,12 +1,11 @@
 package io.deepstream;
 
 import java.net.URI;
-import java.net.URISyntaxException;
 
 class JavaEndpointFactory implements EndpointFactory {
 
     @Override
-    public Endpoint createEndpoint(URI uri, Connection connection) throws URISyntaxException {
+    public Endpoint createEndpoint(URI uri, Connection connection) {
         return new JavaEndpointWebsocket(uri, connection);
     }
 }
