@@ -19,7 +19,7 @@ public interface RecordMergeStrategy {
      * Throw an error if the merge fails, but keep in mind that this only means it will postpone the merge conflict
      * until the next remote/local update.
      *
-     * @param record        The {@link Record}, used to retrieve the local {@link Record#version} and data via {@link Record#get()}
+     * @param record        The {@link Record}, used to retrieve the local {@link Record#version()} and data via {@link Record#get()}
      * @param remoteValue   The remote value on the server
      * @param remoteVersion The remote version on the server, used to find out if the remote is ahead of the local
      * @return The merged value
