@@ -20,15 +20,15 @@ import static org.mockito.Mockito.mock;
 public class RecordSetTest {
 
     private Record record;
-    private ConnectionMock connectionMock;
-    private DeepstreamClientMock deepstreamClientMock;
+    private MockConnection mockConnection;
+    private MockDeepstreamClient deepstreamClientMock;
     private DeepstreamRuntimeErrorHandler errorCallbackMock;
 
     @Before
     public void setUp() throws URISyntaxException, InvalidDeepstreamConfig {
-        this.connectionMock = new ConnectionMock();
+        this.mockConnection = new MockConnection();
         this.errorCallbackMock = mock(DeepstreamRuntimeErrorHandler.class);
-        this.deepstreamClientMock = new DeepstreamClientMock();
+        this.deepstreamClientMock = new MockDeepstreamClient();
         this.deepstreamClientMock.setRuntimeErrorHandler(errorCallbackMock);
         this.deepstreamClientMock.setConnectionState(ConnectionState.OPEN);
 
@@ -38,9 +38,9 @@ public class RecordSetTest {
         options.put("recordReadAckTimeout", "10");
         options.put("recordReadTimeout", "20");
 
-        this.record = new Record("testRecord", new HashMap(), connectionMock, new DeepstreamConfig(options), deepstreamClientMock);
-        record.onMessage(MessageParser.parseMessage(TestUtil.replaceSeperators("R|A|S|testRecord"), deepstreamClientMock));
-        record.onMessage(MessageParser.parseMessage(TestUtil.replaceSeperators("R|R|testRecord|0|{}"), deepstreamClientMock));
+        this.record = new Record("testRecord", new HashMap(), mockConnection, new DeepstreamConfig(options), deepstreamClientMock);
+        record.onMessage(MessageParser.parseMessage(TestUtil.formatMessage("R|A|S|testRecord"), deepstreamClientMock));
+        record.onMessage(MessageParser.parseMessage(TestUtil.formatMessage("R|R|testRecord|0|{}"), deepstreamClientMock));
         Assert.assertEquals(new JsonObject(), record.get());
     }
 
@@ -55,7 +55,7 @@ public class RecordSetTest {
         object.addProperty("firstname", "Wolfram");
         record.set(object);
 
-        Assert.assertEquals(connectionMock.lastSentMessage, TestUtil.replaceSeperators("R|U|testRecord|1|{\"firstname\":\"Wolfram\"}+"));
+        Assert.assertEquals(mockConnection.lastSentMessage, TestUtil.formatMessage("R|U|testRecord|1|{\"firstname\":\"Wolfram\"}+"));
         Assert.assertEquals(object, record.get());
     }
 
@@ -65,21 +65,21 @@ public class RecordSetTest {
 
         record.set("lastname", "Hempel");
 
-        Assert.assertEquals(connectionMock.lastSentMessage, TestUtil.replaceSeperators("R|P|testRecord|2|lastname|SHempel+"));
+        Assert.assertEquals(mockConnection.lastSentMessage, TestUtil.formatMessage("R|P|testRecord|2|lastname|SHempel+"));
         Assert.assertEquals("Hempel", record.get("lastname").getAsString());
     }
 
     @Test
     public void sendsPatchForNestedUpdate() {
         record.set("address[0].street", "randomStreet");
-        Assert.assertEquals(connectionMock.lastSentMessage, TestUtil.replaceSeperators("R|P|testRecord|1|address[0].street|SrandomStreet+"));
+        Assert.assertEquals(mockConnection.lastSentMessage, TestUtil.formatMessage("R|P|testRecord|1|address[0].street|SrandomStreet+"));
         Assert.assertEquals(record.get("address[ 0 ].street").getAsString(), "randomStreet");
     }
 
     @Test
     public void deletesValueWhenReceivingUndefined() {
         record.set("firstname", "Alex");
-        record.onMessage(MessageParser.parseMessage(TestUtil.replaceSeperators("R|P|testRecord|2|firstname|U+"), deepstreamClientMock));
+        record.onMessage(MessageParser.parseMessage(TestUtil.formatMessage("R|P|testRecord|2|firstname|U+"), deepstreamClientMock));
         Object o = record.get("firstname");
         Assert.assertTrue(record.get("firstname").isJsonNull());
     }
@@ -91,7 +91,7 @@ public class RecordSetTest {
         drinks.addProperty("beer", "pilsner");
         drinks.addProperty("coffee", "flat white");
         record.set("drinks", drinks);
-        record.onMessage(MessageParser.parseMessage(TestUtil.replaceSeperators("R|P|testRecord|3|drinks.beer|U+"), deepstreamClientMock));
+        record.onMessage(MessageParser.parseMessage(TestUtil.formatMessage("R|P|testRecord|3|drinks.beer|U+"), deepstreamClientMock));
         Assert.assertTrue(record.get("drinks.beer").isJsonNull());
     }
 
@@ -102,14 +102,14 @@ public class RecordSetTest {
         drinks.add("beer");
         drinks.add("coffee");
         record.set("drinks", drinks);
-        record.onMessage(MessageParser.parseMessage(TestUtil.replaceSeperators("R|P|testRecord|3|drinks[ 0 ]|U+"), deepstreamClientMock));
+        record.onMessage(MessageParser.parseMessage(TestUtil.formatMessage("R|P|testRecord|3|drinks[ 0 ]|U+"), deepstreamClientMock));
         Assert.assertEquals(record.get("drinks[ 0 ]").getAsString(), "coffee");
     }
 
     @Test
     public void deletesNestedArrayValueWhenReceivingUndefined() {
         record.set("address[0].street", "Alex");
-        record.onMessage(MessageParser.parseMessage(TestUtil.replaceSeperators("R|P|testRecord|2|address[ 0 ].street|U+"), deepstreamClientMock));
+        record.onMessage(MessageParser.parseMessage(TestUtil.formatMessage("R|P|testRecord|2|address[ 0 ].street|U+"), deepstreamClientMock));
         Object o = record.get("address[ 0 ].street");
         Assert.assertTrue(record.get("address[ 0 ].street").isJsonNull());
     }
